+++ src/com/kaylerrenslow/armaDialogCreator/control/ControlClassOverridePropertyUpdate.java
@@ -13,30 +13,38 @@ package com.kaylerrenslow.armaDialogCreator.control;
 import org.jetbrains.annotations.NotNull;
 
 /**
- Used when {@link ControlClass#getPropertyUpdateGroup()} is updated via {@link com.kaylerrenslow.armaDialogCreator.util.UpdateListenerGroup#update(Object)}
+ Update triggered by {@link ControlClass#overrideProperty(ControlPropertyLookup)}
 
  @author Kayler
- @since 11/16/16 */
-public class ControlClassPropertyUpdate implements ControlClassUpdate {
+ @since 11/16/2016 */
+public class ControlClassOverridePropertyUpdate implements ControlClassUpdate {
 	private final ControlClass controlClass;
-	private final ControlPropertyUpdate propertyUpdate;
+	private final ControlProperty overidden;
+	private final boolean wasAdded;
 
 	/**
-	 @param controlClass the {@link ControlClass} that was updated
-	 @param propertyUpdate the update
+	 @param controlClass {@link ControlClass} that was updated
+	 @param overridden the {@link ControlProperty} that was overridden
+	 @param wasAdded true if {@link #getOveridden()} was added to {@link ControlClass#getOverriddenProperties()}, false if it was removed
 	 */
-	public ControlClassPropertyUpdate(@NotNull ControlClass controlClass, @NotNull ControlPropertyUpdate propertyUpdate) {
+	public ControlClassOverridePropertyUpdate(@NotNull ControlClass controlClass, @NotNull ControlProperty overridden, boolean wasAdded) {
 		this.controlClass = controlClass;
-		this.propertyUpdate = propertyUpdate;
+		this.overidden = overridden;
+		this.wasAdded = wasAdded;
+	}
+
+	@NotNull
+	public ControlProperty getOveridden() {
+		return overidden;
+	}
+
+	/** @return true if {@link #getOveridden()} was added to {@link ControlClass#getOverriddenProperties()}, false if it was removed */
+	public boolean wasAdded() {
+		return wasAdded;
 	}
 
 	@Override
 	public @NotNull ControlClass getControlClass() {
 		return controlClass;
-	}
-
-	@NotNull
-	public ControlPropertyUpdate getPropertyUpdate() {
-		return propertyUpdate;
 	}
 }