+++ CHANGES.md
@@ -1,9 +1,8 @@
 **Added:**
-* control class update group
+* comments for custom control classes
 
 **Changed:**
 * none
 
 **Fixed:**
-* canvas controls: step labels weren't being displayed properly on small screen sizes (... was inside label text)
-* project init window: fixed "create new project" button size not being big enough for small screen sizes
+* none